third_party/WebKit/LayoutTests/platform/win/virtual/service-worker-navigation-preload-disabled/http/tests/serviceworker/chromium/frame-detached-by-navigation-expected.txt

This is a testharness.js-based test.
PASS navigator.serviceWorker should return null when frame is detached 
Harness: the test ran to completion.
